Move classes to usageanalytics package
