--- analytics/usage/src/main/kotlin/com/malinskiy/marathon/analytics/UsageAnalytics.kt
@@ -1,13 +1,0 @@
-package com.malinskiy.marathon.analytics
-
-import com.malinskiy.marathon.analytics.tracker.EmptyTracker
-import com.malinskiy.marathon.analytics.tracker.GoogleAnalyticsTracker
-import com.malinskiy.marathon.analytics.tracker.Tracker
-
-object UsageAnalytics {
-    var enable = false
-
-    val tracker: Tracker by lazy {
-        if (enable) GoogleAnalyticsTracker() else EmptyTracker()
-    }
-}
