--- analytics/usage/src/main/kotlin/com/malinskiy/marathon/analytics/tracker/Event.kt
@@ -1,6 +1,0 @@
-package com.malinskiy.marathon.analytics.tracker
-
-import com.malinskiy.marathon.analytics.TrackActionType
-
-data class Event(val action: TrackActionType,
-                 val label: String)
