+++ analytics/usage/src/main/kotlin/com/malinskiy/marathon/usageanalytics/Constants.kt
@@ -1,4 +1,4 @@
-package com.malinskiy.marathon.analytics
+package com.malinskiy.marathon.usageanalytics
 
 internal object Constants {
     const val GoogleAnalyticsId = "UA-131132823-1"
