+++ analytics/usage/src/main/kotlin/com/malinskiy/marathon/usageanalytics/UsageAnalytics.kt
@@ -1,0 +1,13 @@
+package com.malinskiy.marathon.usageanalytics
+
+import com.malinskiy.marathon.usageanalytics.tracker.EmptyTracker
+import com.malinskiy.marathon.usageanalytics.tracker.GoogleAnalyticsTracker
+import com.malinskiy.marathon.usageanalytics.tracker.Tracker
+
+object UsageAnalytics {
+    var enable = false
+
+    val tracker: Tracker by lazy {
+        if (enable) GoogleAnalyticsTracker() else EmptyTracker()
+    }
+}
