+++ analytics/usage/src/main/kotlin/com/malinskiy/marathon/usageanalytics/tracker/EmptyTracker.kt
@@ -1,4 +1,4 @@
-package com.malinskiy.marathon.analytics.tracker
+package com.malinskiy.marathon.usageanalytics.tracker
 
 internal class EmptyTracker : Tracker{
     override fun trackEvent(event: Event) {
