+++ analytics/usage/src/main/kotlin/com/malinskiy/marathon/usageanalytics/tracker/Event.kt
@@ -1,0 +1,6 @@
+package com.malinskiy.marathon.usageanalytics.tracker
+
+import com.malinskiy.marathon.usageanalytics.TrackActionType
+
+data class Event(val action: TrackActionType,
+                 val label: String)
