+++ analytics/usage/src/main/kotlin/com/malinskiy/marathon/usageanalytics/tracker/Tracker.kt
@@ -1,4 +1,4 @@
-package com.malinskiy.marathon.analytics.tracker
+package com.malinskiy.marathon.usageanalytics.tracker
 
 interface Tracker{
     fun trackEvent(event: Event)
