+++ cli/src/main/kotlin/com/malinskiy/marathon/cli/ApplicationView.kt
@@ -7,9 +7,9 @@ import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import com.malinskiy.marathon.BuildConfig
 import com.malinskiy.marathon.Marathon
-import com.malinskiy.marathon.analytics.TrackActionType
-import com.malinskiy.marathon.analytics.UsageAnalytics
-import com.malinskiy.marathon.analytics.tracker.Event
+import com.malinskiy.marathon.usageanalytics.TrackActionType
+import com.malinskiy.marathon.usageanalytics.UsageAnalytics
+import com.malinskiy.marathon.usageanalytics.tracker.Event
 import com.malinskiy.marathon.cli.args.MarathonCliConfiguration
 import com.malinskiy.marathon.cli.args.environment.SystemEnvironmentReader
 import com.malinskiy.marathon.cli.config.ConfigFactory
