+++ marathon-gradle-plugin/build.gradle.kts
@@ -21,5 +21,5 @@ dependencies {
     implementation(project(":core"))
     implementation(project(":vendor:vendor-android"))
     implementation(BuildPlugins.androidGradle)
-    implementation(project(":analytics:usage"))
+    api(project(":analytics:usage"))
 }
